+++ ultralytics/models/sam/modules/encoders_.py
