+++ ultralytics/models/sam/modules/utils.py
